requirements.txt
# Dependencias principales del Sistema de Control de Presión
# Fecha: 2025-06-16 15:38:31 UTC

# GUI Framework
PyQt6>=6.4.0

# Configuración
PyYAML>=6.0

# Validación de datos
pydantic>=2.0.0

# Seguridad y autenticación
bcrypt>=4.0.1

# Hardware Raspberry Pi (solo en producción)
RPi.GPIO>=0.7.1; platform_machine=="armv7l" or platform_machine=="aarch64"
gpiozero>=1.6.2; platform_machine=="armv7l" or platform_machine=="aarch64"

# Utilidades
pathlib2>=2.3.7